+++ src/app/components/oneprovider-view-container.js
@@ -23,6 +23,7 @@ import {
   gt,
   and,
   not,
+  or,
 } from 'ember-awesome-macros';
 import safeExec from 'onedata-gui-common/utils/safe-method-execution';
 import { next } from '@ember/runloop';
@@ -33,7 +34,8 @@ import notImplementedIgnore from 'onedata-gui-common/utils/not-implemented-ignor
 import { guidFor } from '@ember/object/internals';
 import isStandaloneGuiOneprovider from 'onedata-gui-common/utils/is-standalone-gui-oneprovider';
 import ChooseDefaultOneprovider from 'onezone-gui/mixins/choose-default-oneprovider';
-import { resolve } from 'rsvp';
+import Version from 'onedata-gui-common/utils/version';
+import computedLastProxyContent from 'onedata-gui-common/utils/computed-last-proxy-content';
 
 const nameComparator = createPropertyComparator('name');
 
@@ -73,12 +75,13 @@ const OneproviderTabItem = EmberObject.extend({
 export default Component.extend(I18n, ChooseDefaultOneprovider, {
   tagName: '',
 
+  media: service(),
+  pointerEvents: service(),
+
   /**
    * @override
    */
   i18nPrefix: 'components.oneproviderViewContainer',
-
-  pointerEvents: service(),
 
   /**
    * View to show in legacy Oneprovider, eg.spaces, transfers, shares
@@ -224,18 +227,19 @@ export default Component.extend(I18n, ChooseDefaultOneprovider, {
 
   validatedOneproviderIdProxy: promise.object(computed(
     'oneproviderId',
-    function validatedOneproviderIdProxy() {
-      const oneproviderId = this.get('oneproviderId');
-      if (oneproviderId) {
-        return resolve(oneproviderId);
+    'minOneproviderRequiredVersion',
+    async function validatedOneproviderIdProxy() {
+      if (this.oneproviderId) {
+        return this.oneproviderId;
       } else {
-        return this.chooseDefaultOneprovider().then(defaultOneprovider => {
-          if (defaultOneprovider) {
-            return resolve(get(defaultOneprovider, 'entityId'));
-          } else {
-            return resolve(null);
-          }
+        const defaultOneprovider = await this.chooseDefaultOneprovider({
+          requiredVersion: this.minOneproviderRequiredVersion,
         });
+        if (defaultOneprovider) {
+          return get(defaultOneprovider, 'entityId');
+        } else {
+          return null;
+        }
       }
     }
   )),
@@ -336,23 +340,58 @@ export default Component.extend(I18n, ChooseDefaultOneprovider, {
     }
   ),
 
+  showAllVersionsOld: computed(
+    'selectedProvider',
+    'hasOneproviderInRequiredVersion',
+    function showAllVersionsOld() {
+      return !this.selectedProvider && this.hasOneproviderInRequiredVersion === false;
+    }
+  ),
+
+  isNoSelectedProviderErrorShown: and(
+    'oneproviderViewProxy.isFulfilled',
+    not('selectedProvider'),
+  ),
+
+  isProvidersTabBarForced: or(
+    'showAllOfflineInfo',
+    'showAllVersionsOld',
+    'isNoSelectedProviderErrorShown'
+  ),
+
+  effIsTabBarCollapsed: and(
+    'isTabBarCollapsed',
+    not('media.isMobile'),
+    not('isProvidersTabBarForced'),
+  ),
+
   hasSupport: notEmpty('providers'),
 
+  // FIXME: trzeba sprawdzić co się dzieje jak dochodzi nowy provider, czy nie miga
   /**
    * @type {ComputedProperty<PromiseObject<Array<Model.Provider>>>}
    */
   initialProvidersListProxy: promise.object(
-    computed('space.providerList', function initialProvidersListProxy() {
-      return this.get('space').getRelation('providerList')
-        .then(providerList => get(providerList, 'list'))
-        .then(list => sortedOneprovidersList(list.toArray()));
+    computed('space.providerList', async function initialProvidersListProxy() {
+      const providerList = await this.space.getRelation('providerList');
+      const providers = await get(providerList, 'list');
+      return sortedOneprovidersList(providers.toArray());
     })
   ),
+
+  viewRequiredDataProxy: promise.object(computed(
+    async function viewRequiredDataProxy() {
+      await this.initialProvidersListProxy;
+      await this.hasOneproviderInRequiredVersionProxy;
+    }
+  )),
+
+  // FIXME: hasOneproviderInRequiredVersionProxy powinno być robione dopiero po resolve initialProvidersListProxy
 
   isEmbeddableOneproviderProxy: promise.object(
     computed(
       'initialProvidersListProxy',
-      'selectedProvider.versionProxy',
+      'selectedProvider.version',
       function isEmbeddableOneproviderProxy() {
         return this.get('initialProvidersListProxy').then(() => {
           const selectedProvider = this.get('selectedProvider');
@@ -360,9 +399,8 @@ export default Component.extend(I18n, ChooseDefaultOneprovider, {
             if (get(selectedProvider, 'id') === 'overview') {
               return selectedProvider;
             }
-            return get(selectedProvider, 'versionProxy').then(version => {
-              return !isStandaloneGuiOneprovider(version);
-            });
+            const version = get(selectedProvider, 'version');
+            return !isStandaloneGuiOneprovider(version);
           } else {
             return null;
           }
@@ -371,6 +409,36 @@ export default Component.extend(I18n, ChooseDefaultOneprovider, {
   ),
 
   isEmbeddableOneprovider: reads('isEmbeddableOneproviderProxy.content'),
+
+  // FIXME: move
+  /**
+   * Oneprovider version string or null. Eg. `21.02.1`.
+   * @type {string | null}
+   */
+  minOneproviderRequiredVersion: null,
+
+  hasOneproviderInRequiredVersionProxy: promise.object(computed(
+    'minOneproviderRequiredVersion',
+    'providers.@each.version',
+    async function hasOneproviderInRequiredVersionProxy() {
+      if (!this.providers?.length) {
+        return false;
+      }
+      const requiredVersion = this.minOneproviderRequiredVersion;
+      if (!requiredVersion) {
+        return Boolean(this.providers?.length);
+      }
+
+      return this.providers.some((provider) => {
+        const providerVersion = get(provider, 'version');
+        return Version.isRequiredVersion(providerVersion, requiredVersion);
+      });
+    }
+  )),
+
+  hasOneproviderInRequiredVersion: computedLastProxyContent(
+    'hasOneproviderInRequiredVersionProxy'
+  ),
 
   /**
    * When there is no Oneprovider selected (or there is no Oneprovider at all)
@@ -390,18 +458,12 @@ export default Component.extend(I18n, ChooseDefaultOneprovider, {
   init() {
     this._super(...arguments);
 
-    const {
-      initialProvidersListProxy,
-      isOneproviderViewLocal,
-    } = this.getProperties('initialProvidersListProxy', 'isOneproviderViewLocal');
-
-    initialProvidersListProxy.then(list => {
-      const oneproviderId = this.get('oneproviderId');
-      if (!oneproviderId) {
-        return this.selectDefaultProvider(list);
+    this.initialProvidersListProxy.then(list => {
+      if (!this.oneproviderId) {
+        this.selectDefaultProvider(list);
       }
     });
-    if (!isOneproviderViewLocal) {
+    if (!this.isOneproviderViewLocal) {
       next(() => {
         safeExec(this, 'set', 'pointerEvents.pointerNoneToMainContent', true);
       });
@@ -415,16 +477,18 @@ export default Component.extend(I18n, ChooseDefaultOneprovider, {
     });
   },
 
-  selectDefaultProvider(providers = this.get('providers')) {
-    return this.chooseDefaultOneprovider(providers).then(defaultProvider => {
-      if (defaultProvider) {
-        const providerId = get(defaultProvider, 'entityId');
-        this.get('oneproviderIdChanged')(
-          providerId,
-          true,
-        );
-      }
+  async selectDefaultProvider(providers = this.providers) {
+    const defaultProvider = await this.chooseDefaultOneprovider({
+      providers,
+      requiredVersion: this.minOneproviderRequiredVersion,
     });
+    if (defaultProvider) {
+      const providerId = get(defaultProvider, 'entityId');
+      this.oneproviderIdChanged(
+        providerId,
+        true,
+      );
+    }
   },
 
   actions: {