+++ src/app/locales/en/components/content-spaces-old-oneprovider.js
@@ -1,0 +1,9 @@
+export default {
+  header: 'Not supported by older providers',
+  text: 'This functionality is supported since Oneprovider version {{requiredVersion}}.',
+  info: {
+    allOld: 'All providers in this space are in older versions and must be upgraded for the new features to be enabled.',
+    allNewOffline: 'Currently, all the providers that have a sufficient version are offline.',
+    selectedOld: 'This provider is in version $version and must be upgraded for the new features to be enabled.',
+  },
+};