+++ src/app/mixins/choose-default-oneprovider.js
@@ -3,16 +3,17 @@
  * Oneprovider if not chosen yet.
  *
  * @author Jakub Liput
- * @copyright (C) 2020 ACK CYFRONET AGH
+ * @copyright (C) 2020-2023 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 
 import Mixin from '@ember/object/mixin';
-import { resolve, allSettled } from 'rsvp';
+import { get } from '@ember/object';
 import isStandaloneGuiOneprovider from 'onedata-gui-common/utils/is-standalone-gui-oneprovider';
 import createPropertyComparator from 'onedata-gui-common/utils/create-property-comparator';
 import { assert } from '@ember/debug';
 import globals from 'onedata-gui-common/utils/globals';
+import Version from 'onedata-gui-common/utils/version';
 
 const storageOneproviderIdKey = 'chooseDefaultOneproviderMixin.oneproviderId';
 
@@ -21,7 +22,6 @@ export function getStorageOneproviderKey(spaceId) {
 }
 
 export default Mixin.create({
-
   getStorageOneproviderKey,
 
   getBrowserDefaultOneproviderId(spaceId) {
@@ -36,45 +36,59 @@ export default Mixin.create({
   /**
    * Find Oneprovider that should be currently chosen if saved default Oneprovider
    * is not available
-   * @param {Array<Models.Provider>} onlineOneproviders
+   * @param {Array<Models.Provider>} oneproviders
    * @returns {Models.Provider}
    */
-  findCurrentDefaultOneprovider(onlineOneproviders) {
+  findCurrentDefaultOneprovider(oneproviders) {
     assert(
-      'findCurrentDefaultOneprovider: onlineOneproviders should be not null/undefined',
-      onlineOneproviders
+      'findCurrentDefaultOneprovider: oneproviders should be not null/undefined',
+      oneproviders
     );
-    const sortedOnlineOneproviders = [...onlineOneproviders.toArray()].sort(
+    const sortedApplicableOneproviders = [...oneproviders.toArray()].sort(
       createPropertyComparator('name')
     );
-    return allSettled(sortedOnlineOneproviders.mapBy('versionProxy'))
-      .then(onlineVersionsResult => {
-        const oneprovider = sortedOnlineOneproviders.objectAt(
-          onlineVersionsResult.findIndex(({ state, value }) =>
-            state === 'fulfilled' && !isStandaloneGuiOneprovider(value)
-          )
-        );
-        if (oneprovider) {
-          return oneprovider;
-        } else {
-          return sortedOnlineOneproviders[0];
-        }
-      });
+    if (!sortedApplicableOneproviders.length) {
+      return null;
+    }
+    // prefer embeddable (20.02+) providers as default
+    const oneprovider = sortedApplicableOneproviders.find(provider => {
+      const version = get(provider, 'version');
+      return !isStandaloneGuiOneprovider(version);
+    });
+    if (oneprovider) {
+      return oneprovider;
+    } else {
+      return sortedApplicableOneproviders[0];
+    }
   },
 
-  chooseDefaultOneprovider(
-    oneproviders = (this.get('providers') || []),
-    spaceId = this.get('space.entityId')
-  ) {
-    if (!oneproviders) {
-      return resolve(null);
+  async chooseDefaultOneprovider({
+    providers = (this.providers ?? []),
+    spaceId = this.get('space.entityId'),
+    requiredVersion,
+  } = {}) {
+    if (!providers) {
+      return null;
     }
     const defaultId = this.getBrowserDefaultOneproviderId(spaceId);
-    const onlineOneproviders = oneproviders.filterBy('online');
+    let applicableProviders = providers.filter(provider => get(provider, 'online'));
+    if (requiredVersion) {
+      applicableProviders = providers.filter(provider => {
+        try {
+          const providerVersion = get(provider, 'version');
+          return Version.isRequiredVersion(providerVersion, requiredVersion);
+        } catch {
+          return false;
+        }
+      });
+    }
     const savedDefaultOneprovider = (
-      defaultId && onlineOneproviders.findBy('entityId', defaultId)
+      defaultId && applicableProviders.findBy('entityId', defaultId)
     );
-    return savedDefaultOneprovider && resolve(savedDefaultOneprovider) ||
-      this.findCurrentDefaultOneprovider(onlineOneproviders);
+    // FIXME: debug
+    console.log('savedDefaultOneprovider', savedDefaultOneprovider);
+    console.log('applicable', applicableProviders);
+    return savedDefaultOneprovider ||
+      this.findCurrentDefaultOneprovider(applicableProviders);
   },
 });